+++ src/p2p/net_node.inl
@@ -52,7 +52,6 @@
 #include "common/pruning.h"
 #include "net/error.h"
 #include "net/net_helper.h"
-#include "math_helper.h"
 #include "misc_log_ex.h"
 #include "p2p_protocol_defs.h"
 #include "crypto/crypto.h"
